Удалите когда-нибудь этот коммит, очень тупые изменения, комменты авто, логгера, авто вычитаны и проверены
