+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/pedropathingauto/blue/Auto12ArtefactsShortBlue.java
@@ -27,8 +27,8 @@ public class Auto12ArtefactsShortBlue extends LinearOpMode {
     private Timer actionTimer;
     public Pose currentPose; // Current pose of the robot
 
-    Intake in;
-    Shooter sh;
+//    Intake in;
+//    Shooter sh;
 
     @Override
     public void runOpMode() {
@@ -39,8 +39,8 @@ public class Auto12ArtefactsShortBlue extends LinearOpMode {
         actionTimer = new Timer();
         actionTimer.resetTimer();
 
-        in = new Intake(this);
-        sh = new Shooter(this);
+//        in = new Intake(this);
+//        sh = new Shooter(this);
         Logger lg = new Logger("pospos");
 
         Telemetry dash = FtcDashboard.getInstance().getTelemetry();
@@ -50,9 +50,9 @@ public class Auto12ArtefactsShortBlue extends LinearOpMode {
 
         paths = new Paths(follower); // Build paths
 
-        sh.closeTunnel();
-        sh.setShortThrowMode();
-        sh.setVelocityTarget(Shooter.VELOCITY_FOR_SHORT_THROW);
+//        sh.closeTunnel();
+//        sh.setShortThrowMode();
+//        sh.setVelocityTarget(Shooter.VELOCITY_FOR_SHORT_THROW);
 
         waitForStart();
         while (opModeIsActive()) {
@@ -61,14 +61,14 @@ public class Auto12ArtefactsShortBlue extends LinearOpMode {
             currentPose = follower.getPose(); // Update the current pose
 
 
-            sh.threeArtefactsShooting();
+           // sh.threeArtefactsShooting();
 
             // Log values to Panels and Driver Station
             t.addData("Path State", pathState);
             t.addData("X", follower.getPose().getX());
             t.addData("Y", follower.getPose().getY());
             t.addData("Heading", follower.getPose().getHeading());
-            t.addData("Shooter Velocity", sh.getVelocityRPS());
+           // t.addData("Shooter Velocity", sh.getVelocityRPS());
             t.update();
         }
         lg.writePose(Alliance.BLUE, follower.getPose().getX(), follower.getPose().getY(), follower.getPose().getHeading());
@@ -209,16 +209,17 @@ public class Auto12ArtefactsShortBlue extends LinearOpMode {
     public void autonomousPathUpdate() {
         switch (pathState) {
             case 0:
-                sh.closeTunnel();
-                sh.shootByVelocity();
-                in.rotateIn();
+//                sh.closeTunnel();
+//                sh.shootByVelocity();
+//                in.rotateIn();
                 follower.followPath(paths.PathScoring);
                 setPathState(1);
                 break;
 
             case 1:
-                if (!sh.isSpinUp()) break;
-                sh.openTunnel();
+//                if (!sh.isSpinUp()) break;
+//                sh.openTunnel();
+                if(follower.isBusy()) break;
                 setPathState(2);
                 break;
 
@@ -230,8 +231,8 @@ public class Auto12ArtefactsShortBlue extends LinearOpMode {
 
             case 4:
                 if (!follower.isBusy()) {
-                    in.rotateIn();
-                    sh.closeTunnel();
+//                    in.rotateIn();
+//                    sh.closeTunnel();
                     follower.followPath(paths.PathIntakingArtifacts, true);
                     setPathState(5);
                 }
@@ -245,72 +246,79 @@ public class Auto12ArtefactsShortBlue extends LinearOpMode {
                 break;
 
             case 6:
-                if (!sh.isSpinUp()) break;
-                sh.openTunnel();
+//                if (!sh.isSpinUp()) break;
+//                sh.openTunnel();
+                if(follower.isBusy()) break;
                 setPathState(7);
                 break;
 
             case 7:
                 if (follower.isBusy() || actionTimer.getElapsedTime() < 4000) break;
                 follower.followPath(paths.PathSecondPresentArtefacts, true);
-                sh.closeTunnel();
-                in.rotateIn();
+//                sh.closeTunnel();
+//                in.rotateIn();
                 setPathState(8);
                 break;
 
             case 8:
-                if (!follower.isBusy())
+                if (!follower.isBusy()) {
                     follower.followPath(paths.PathSecondIntakingArtefacts, true);
-                setPathState(9);
+                    setPathState(9);
+                }
                 break;
 
             case 9:
-                if (!follower.isBusy())
+                if (!follower.isBusy()) {
                     follower.followPath(paths.PathThirdScoring, true);
-                setPathState(10);
+                    setPathState(10);
+                }
                 break;
 
             case 10:
                 if (follower.isBusy() || actionTimer.getElapsedTime() < 4000) break;
-                sh.openTunnel();
+               // sh.openTunnel();
                 setPathState(11);
                 break;
 
             case 11:
-                sh.closeTunnel();
-                if (!follower.isBusy())
+               // sh.closeTunnel();
+                if (!follower.isBusy()) {
                     follower.followPath(paths.PathThirdPresentArtefacts, true);
-                setPathState(12);
+                    setPathState(12);
+                }
                 break;
 
             case 12:
-                in.rotateIn();
-                if (!follower.isBusy())
+              //  in.rotateIn();
+                if (!follower.isBusy()) {
                     follower.followPath(paths.PathThirdIntakingArtefacts, true);
-                setPathState(13);
+                    setPathState(13);
+                }
                 break;
 
             case 13:
-                if (!follower.isBusy())
+                if (!follower.isBusy()) {
                     follower.followPath(paths.PathFourScoring, true);
-                setPathState(14);
+                    setPathState(14);
+                }
                 break;
 
             case 14:
-                if (!follower.isBusy())
+                if (!follower.isBusy()) {
                     follower.followPath(paths.PathThirdIntakingArtefacts, true);
-                setPathState(15);
+                    setPathState(15);
+                }
                 break;
 
             case 15:
-                if (!sh.isSpinUp()) break;
-                sh.openTunnel();
+//                if (!sh.isSpinUp()) break;
+//                sh.openTunnel();
                 setPathState(16);
                 break;
 
             case 16:
                 if (follower.isBusy() || actionTimer.getElapsedTime() < 4000) break;
-                in.rotateStop();
+              //  in.rotateStop();
                 follower.followPath(paths.PathLeaving);
                 setPathState(-100);
                 break;
