--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/pedropathingauto/blue/Auto3ArtefactsLongBlue.java
@@ -1,124 +1,0 @@
-package org.firstinspires.ftc.teamcode.opmodes.auto.pedropathingauto.blue;
-
-import com.acmerobotics.dashboard.FtcDashboard;
-import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
-import com.pedropathing.follower.Follower;
-import com.pedropathing.geometry.BezierLine;
-import com.pedropathing.geometry.Pose;
-import com.pedropathing.paths.PathChain;
-import com.pedropathing.util.Timer;
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.modules.Intake;
-import org.firstinspires.ftc.teamcode.modules.Shooter;
-import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
-import org.firstinspires.ftc.teamcode.util.Alliance;
-import org.firstinspires.ftc.teamcode.util.Logger;
-
-@Autonomous(name = "Auto 3 Long BLUE", group = "1")
-public class Auto3ArtefactsLongBlue extends LinearOpMode {
-    public Follower follower; // Pedro Pathing follower instance
-    private int pathState; // Current autonomous path state (state machine)
-    private Auto12ArtefactsLongBlue.Paths paths; // Paths defined in the Paths class
-    private Timer pathTimer;
-    private Timer actionTimer;
-    public Pose currentPose; // Current pose of the robot
-
-    Intake in;
-    Shooter sh;
-
-    @Override
-    public void runOpMode() {
-        pathTimer = new Timer();
-        Timer opmodeTimer = new Timer();
-        opmodeTimer.resetTimer();
-
-        actionTimer = new Timer();
-        actionTimer.resetTimer();
-
-        in = new Intake(this);
-        sh = new Shooter(this);
-        Logger lg = new Logger("pospos");
-
-        Telemetry dash = FtcDashboard.getInstance().getTelemetry();
-        Telemetry t = new MultipleTelemetry(telemetry, dash);
-        follower = Constants.createFollower(hardwareMap);
-        follower.setStartingPose(new Pose(60, 8, Math.toRadians(90)));
-
-        paths = new Auto12ArtefactsLongBlue.Paths(follower); // Build paths
-
-        sh.closeTunnel();
-        sh.setLongThrowMode();
-        sh.setVelocityTarget(Shooter.VELOCITY_FOR_LONG_THROW);
-
-        waitForStart();
-        while (opModeIsActive()) {
-            follower.update();
-            autonomousPathUpdate();
-            currentPose = follower.getPose();
-
-            paths = new Auto12ArtefactsLongBlue.Paths(follower);
-
-            t.addData("Path State", pathState);
-            t.addData("X", follower.getPose().getX());
-            t.addData("Y", follower.getPose().getY());
-            t.addData("Heading", follower.getPose().getHeading());
-            t.addData("Shooter Velocity", sh.getVelocityRPS());
-            t.update();
-        }
-        lg.writePose(Alliance.BLUE, follower.getPose().getX(), follower.getPose().getY(), follower.getPose().getHeading());
-        lg.fileClose();
-    }
-
-    public static class Paths {
-        public final PathChain PathScoring;
-
-        public Paths(Follower follower) {
-            PathScoring = follower.pathBuilder().addPath(
-                            new BezierLine(
-                                    new Pose(60, 8),
-
-                                    new Pose(60, 15)
-                            )
-                    ).setLinearHeadingInterpolation(Math.toRadians(90), Math.toRadians(110))
-
-                    .build();
-        }
-    }
-
-
-    public void autonomousPathUpdate() {
-        switch (pathState) {
-            case 0:
-                sh.closeTunnel();
-                sh.shootByVelocity();
-                in.rotateIn();
-                follower.followPath(paths.PathScoring);
-                setPathState(1);
-                break;
-
-            case 1:
-                if (!sh.isSpinUp()) break;
-                sh.openTunnel();
-                setPathState(2);
-                break;
-
-            case 2:
-                if (follower.isBusy() || actionTimer.getElapsedTime() < 4000) break;
-                follower.followPath(paths.PathToPresetArtifacts);
-                setPathState(-100);
-                break;
-
-        }
-    }
-
-    public void setPathState(int pState) {
-        pathState = pState;
-        pathTimer.resetTimer();
-        actionTimer.resetTimer();
-    }
-}
-
-
