+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/pedropathingauto/blue/Auto3ArtifactsShortBlue.java
@@ -29,8 +29,8 @@ public class Auto3ArtifactsShortBlue extends LinearOpMode {
     public Pose currentPose; // Current pose of the robot
   //  Turret tt;
 
-    Intake in;
-    Shooter sh;
+//    Intake in;
+//    Shooter sh;
 
     @Override
     public void runOpMode() {
@@ -41,8 +41,8 @@ public class Auto3ArtifactsShortBlue extends LinearOpMode {
         actionTimer = new Timer();
         actionTimer.resetTimer();
 
-        in = new Intake(this);
-        sh = new Shooter(this);
+//        in = new Intake(this);
+//        sh = new Shooter(this);
       //  tt = new Turret(this);
         Logger lg = new Logger("pospos");
 
@@ -53,10 +53,10 @@ public class Auto3ArtifactsShortBlue extends LinearOpMode {
 
         paths = new Paths(follower); // Build paths
 
-        sh.closeTunnel();
-        sh.setLongThrowMode();
-     //   tt.turretRegulator.start();
-        sh.setVelocityTarget(Shooter.VELOCITY_FOR_SHORT_THROW);
+//        sh.closeTunnel();
+//        sh.setLongThrowMode();
+//        tt.turretRegulator.start();
+//        sh.setVelocityTarget(Shooter.VELOCITY_FOR_SHORT_THROW);
 
         waitForStart();
         while (opModeIsActive()) {
@@ -70,7 +70,7 @@ public class Auto3ArtifactsShortBlue extends LinearOpMode {
             t.addData("X", follower.getPose().getX());
             t.addData("Y", follower.getPose().getY());
             t.addData("Heading", follower.getPose().getHeading());
-            t.addData("Shooter Velo", sh.getVelocityRPS());
+           // t.addData("Shooter Velo", sh.getVelocityRPS());
             t.update();
         }
         lg.writePose(Alliance.BLUE, follower.getPose().getX(), follower.getPose().getY(), follower.getPose().getHeading());
@@ -110,28 +110,29 @@ public class Auto3ArtifactsShortBlue extends LinearOpMode {
         switch (pathState) {
             case 0:
                 follower.followPath(paths.PathScoring);
-                sh.closeTunnel();
-                sh.shootByVelocity();
-                in.rotateIn();
+//                sh.closeTunnel();
+//                sh.shootByVelocity();
+//                in.rotateIn();
                 setPathState(1);
                 break;
             case 1:
-                if (!sh.isSpinUp() || follower.isBusy()) break;
-                //while (!sh.isSpinUp() || follower.isBusy());
-                sh.openTunnel();
+                if(follower.isBusy()) break;
+//                if (!sh.isSpinUp() || follower.isBusy()) break;
+//                while (!sh.isSpinUp() || follower.isBusy());
+//                sh.openTunnel();
                 setPathState(2);
                 break;
             case 2:
                 if(follower.isBusy() || actionTimer.getElapsedTime() < 4000) break;
                 // while (follower.isBusy() || actionTimer.getElapsedTime() < 4000) ;
-                sh.closeTunnel();
+                //sh.closeTunnel();
                 follower.followPath(paths.PathLeaving);
-                sh.shootStop();
-                in.rotateStop();
+//                sh.shootStop();
+//                in.rotateStop();
                 setPathState(3);
                 break;
             case 3:
-                if (!follower.isBusy() && !Shooter.isTunnelOpen) {
+                if (!follower.isBusy()){ //&& !Shooter.isTunnelOpen) {
                     setPathState(-100);
                 }
                 break;
