+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/pedropathingauto/blue/Auto6ArtifactsShortBlue.java
@@ -8,7 +8,6 @@ import com.bylazar.telemetry.PanelsTelemetry;
 
 import org.firstinspires.ftc.teamcode.modules.Intake;
 import org.firstinspires.ftc.teamcode.modules.Shooter;
-import org.firstinspires.ftc.teamcode.modules.Transfer;
 import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
 
 import com.pedropathing.geometry.BezierLine;
@@ -28,7 +27,6 @@ public class Auto6ArtifactsShortBlue extends LinearOpMode {
     private Timer actionTimer;
     //private Turret tt;
     public Pose currentPose; // Current pose of the robot
-    public Transfer tr;
 
     Intake in;
     Shooter sh;
@@ -39,9 +37,8 @@ public class Auto6ArtifactsShortBlue extends LinearOpMode {
         Timer opmodeTimer = new Timer();
         opmodeTimer.resetTimer();
 
-        tr = new Transfer(this);
         in = new Intake(this);
-        sh = new Shooter(this, follower, tr);
+        sh = new Shooter(this);
         //tt = new Turret(this);
 
         // Panels Telemetry instance
@@ -55,7 +52,7 @@ public class Auto6ArtifactsShortBlue extends LinearOpMode {
         panelsTelemetry.debug("Status", "Initialized");
         panelsTelemetry.update(telemetry);
 
-      //  sh.closeTunnel();
+        sh.closeTunnel();
         sh.setShortThrowMode();
 
         waitForStart();
@@ -139,13 +136,14 @@ public class Auto6ArtifactsShortBlue extends LinearOpMode {
     public void autonomousPathUpdate() {
         switch (pathState) {
             case 0:
-               sh.setVelocityTarget(Shooter.VELOCITY_FOR_SHORT_THROW);
+                sh.setVelocityTarget(Shooter.VELOCITY_FOR_SHORT_THROW);
                 sh.shootByVelocity();
                 in.rotateIn();
                 setPathState(1);
                 break;
             case 1:
-                if ( !follower.isBusy()) { //sh.isSpinUp() &&
+                if(!follower.isBusy()){
+              //  if (sh.isSpinUp() && !follower.isBusy()) {
                     follower.followPath(paths.PathFirstScoring, true);
                     setPathState(2);
                 }
@@ -157,7 +155,8 @@ public class Auto6ArtifactsShortBlue extends LinearOpMode {
                 }
                 break;
             case 3:
-                if (!follower.isBusy()) {//!Shooter.isTunnelOpen &&
+                if(!follower.isBusy()){
+                //if (!Shooter.isTunnelOpen && !follower.isBusy()) {
                     follower.followPath(paths.PathToPresetArtifacts, true);
                     setPathState(4);
                 }
@@ -181,7 +180,8 @@ public class Auto6ArtifactsShortBlue extends LinearOpMode {
                 }
                 break;
             case 7:
-                if(!follower.isBusy() ){ //&& !Shooter.isTunnelOpen
+                if(!follower.isBusy()){
+                //if(!follower.isBusy() && !Shooter.isTunnelOpen){
                     follower.followPath(paths.PathLeaving);
                     setPathState(-100);
                 }
