+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/pedropathingauto/blue/far/Auto3ArtefactsLongBlue.java
@@ -2,7 +2,6 @@ package org.firstinspires.ftc.teamcode.opmodes.auto.pedropathingauto.blue.far;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
-import com.bylazar.configurables.annotations.Configurable;
 import com.pedropathing.follower.Follower;
 import com.pedropathing.geometry.BezierLine;
 import com.pedropathing.geometry.Pose;
@@ -12,25 +11,21 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.modules.Intake;
-import org.firstinspires.ftc.teamcode.modules.Shooter;
 import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
 import org.firstinspires.ftc.teamcode.util.Alliance;
 import org.firstinspires.ftc.teamcode.util.Logger;
 
-@Autonomous(name = "3 Artifacts Long Auto BLUE", group = "Autonomous")
-@Configurable // Panels
-public class Auto3ArtifactsLongBlue extends LinearOpMode {
+@Autonomous(name = "Auto 3 Long BLUE", group = "1")
+public class Auto3ArtefactsLongBlue extends LinearOpMode {
     public Follower follower; // Pedro Pathing follower instance
     private int pathState; // Current autonomous path state (state machine)
-    private Paths paths; // Paths defined in the Paths class
+    private Auto12ArtefactsLongBlue.Paths paths; // Paths defined in the Paths class
     private Timer pathTimer;
     private Timer actionTimer;
     public Pose currentPose; // Current pose of the robot
- //   Turret tt;
 
-    Intake in;
-    Shooter sh;
+  //  Intake in;
+  //  Shooter sh;
 
     @Override
     public void runOpMode() {
@@ -41,99 +36,88 @@ public class Auto3ArtifactsLongBlue extends LinearOpMode {
         actionTimer = new Timer();
         actionTimer.resetTimer();
 
-        in = new Intake(this);
-        sh = new Shooter(this);
-      //  tt = new Turret(this);
+//        in = new Intake(this);
+//        sh = new Shooter(this);
         Logger lg = new Logger("pospos");
 
         Telemetry dash = FtcDashboard.getInstance().getTelemetry();
         Telemetry t = new MultipleTelemetry(telemetry, dash);
         follower = Constants.createFollower(hardwareMap);
-        follower.setStartingPose(new Pose(56, 8, Math.toRadians(90)));//22,124
+        follower.setStartingPose(new Pose(60, 8, Math.toRadians(90)));
 
-        paths = new Paths(follower); // Build paths
+        paths = new Auto12ArtefactsLongBlue.Paths(follower); // Build paths
 
-        sh.closeTunnel();
-        sh.setLongThrowMode();
-      //  tt.turretRegulator.start();
-        sh.setVelocityTarget(Shooter.VELOCITY_FOR_LONG_THROW);
+//        sh.closeTunnel();
+//        sh.setLongThrowMode();
+//        sh.setVelocityTarget(Shooter.VELOCITY_FOR_LONG_THROW);
 
         waitForStart();
         while (opModeIsActive()) {
-            follower.update(); // Update Pedro Pathing
-            autonomousPathUpdate(); // Update autonomous state machine
-            currentPose = follower.getPose(); // Update the current pose
+            follower.update();
+            autonomousPathUpdate();
+            currentPose = follower.getPose();
 
+            paths = new Auto12ArtefactsLongBlue.Paths(follower);
 
-            sh.threeArtefactsShooting();
-
-            // Log values to Panels and Driver Station
             t.addData("Path State", pathState);
             t.addData("X", follower.getPose().getX());
             t.addData("Y", follower.getPose().getY());
             t.addData("Heading", follower.getPose().getHeading());
-            t.addData("Shooter Velocity", sh.getVelocityRPS());
+         //   t.addData("Shooter Velocity", sh.getVelocityRPS());
             t.update();
         }
         lg.writePose(Alliance.BLUE, follower.getPose().getX(), follower.getPose().getY(), follower.getPose().getHeading());
         lg.fileClose();
-     //   tt.turretRegulator.interrupt();
     }
 
-
     public static class Paths {
-        public final PathChain PathLeaving;
+        public final PathChain PathScoring;
 
         public Paths(Follower follower) {
-            PathLeaving = follower.pathBuilder().addPath(
+            PathScoring = follower.pathBuilder().addPath(
                             new BezierLine(
-                                    new Pose(56, 8),
+                                    new Pose(60, 8),
 
-                                    new Pose(36, 8)
+                                    new Pose(60, 15)
                             )
-                    ).setConstantHeadingInterpolation(Math.toRadians(110))
+                    ).setLinearHeadingInterpolation(Math.toRadians(90), Math.toRadians(110))
 
                     .build();
         }
-
     }
 
 
     public void autonomousPathUpdate() {
         switch (pathState) {
             case 0:
-                sh.closeTunnel();
-                sh.shootByVelocity();
-                in.rotateIn();
+//                sh.closeTunnel();
+//                sh.shootByVelocity();
+//                in.rotateIn();
+                follower.followPath(paths.PathScoring);
                 setPathState(1);
                 break;
+
             case 1:
-                while (!sh.isSpinUp() || follower.isBusy()) ;
-
-                sh.openTunnel();
-              //  sh.threeArtefactsShooting();
-              setPathState(2);
+                if(follower.isBusy()) break;
+//                if (!sh.isSpinUp()) break;
+//                sh.openTunnel();
+                setPathState(2);
                 break;
+
             case 2:
-                while (follower.isBusy() || actionTimer.getElapsedTime() < 4000) ;
-                sh.closeTunnel();
-                follower.followPath(paths.PathLeaving);
-                sh.shootStop();
-                in.rotateStop();
-              //  tt.turnByTarget(0);
-                setPathState(3);
+                if (follower.isBusy() || actionTimer.getElapsedTime() < 4000) break;
+                follower.followPath(paths.PathToPresetArtifacts);
+                setPathState(-100);
                 break;
-            case 3:
-                if (!follower.isBusy() && !Shooter.isTunnelOpen ) {
-                    setPathState(-100);
-                }
-                break;
-        }
 
+        }
     }
 
     public void setPathState(int pState) {
         pathState = pState;
         pathTimer.resetTimer();
+        actionTimer.resetTimer();
     }
 }
+
+
